feat: interfaces now supplied via configuration file
