+++ go.mod
@@ -3,6 +3,10 @@ module github.com/cosmos/cosmos-proto
 go 1.16
 
 require (
+	github.com/cosmos/cosmos-sdk v0.44.0
 	github.com/stretchr/testify v1.7.0
 	google.golang.org/protobuf v1.27.0
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
+
+replace github.com/gogo/protobuf => github.com/regen-network/protobuf v1.3.3-alpha.regen.1
