+++ testpb/1.pulsar.go
@@ -3,6 +3,7 @@ package testpb
 import (
 	fmt "fmt"
 	_ "github.com/cosmos/cosmos-proto"
+	authz "github.com/cosmos/cosmos-sdk/x/authz"
 	types "github.com/cosmos/cosmos-sdk/x/gov/types"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoiface "google.golang.org/protobuf/runtime/protoiface"
@@ -1106,6 +1107,7 @@ var (
 	md_Foo     protoreflect.MessageDescriptor
 	fd_Foo_bro protoreflect.FieldDescriptor
 	fd_Foo_bar protoreflect.FieldDescriptor
+	fd_Foo_baz protoreflect.FieldDescriptor
 )
 
 func init() {
@@ -1113,6 +1115,7 @@ func init() {
 	md_Foo = File_testpb_1_proto.Messages().ByName("Foo")
 	fd_Foo_bro = md_Foo.Fields().ByName("bro")
 	fd_Foo_bar = md_Foo.Fields().ByName("bar")
+	fd_Foo_baz = md_Foo.Fields().ByName("baz")
 }
 
 var _ protoreflect.Message = (*fastReflection_Foo)(nil)
@@ -1192,6 +1195,12 @@ func (x *fastReflection_Foo) Range(f func(protoreflect.FieldDescriptor, protoref
 			return
 		}
 	}
+	if x.Baz != nil {
+		value := protoreflect.ValueOfMessage(x.Baz.ProtoReflect())
+		if !f(fd_Foo_baz, value) {
+			return
+		}
+	}
 }
 
 // Has reports whether a field is populated.
@@ -1211,6 +1220,8 @@ func (x *fastReflection_Foo) Has(fd protoreflect.FieldDescriptor) bool {
 		return x.Bro != ""
 	case "Foo.bar":
 		return x.Bar != nil
+	case "Foo.baz":
+		return x.Baz != nil
 	default:
 		if fd.IsExtension() {
 			panic(fmt.Errorf("proto3 declared messages do not support extensions: Foo"))
@@ -1231,6 +1242,8 @@ func (x *fastReflection_Foo) Clear(fd protoreflect.FieldDescriptor) {
 		x.Bro = ""
 	case "Foo.bar":
 		x.Bar = nil
+	case "Foo.baz":
+		x.Baz = nil
 	default:
 		if fd.IsExtension() {
 			panic(fmt.Errorf("proto3 declared messages do not support extensions: Foo"))
@@ -1252,6 +1265,9 @@ func (x *fastReflection_Foo) Get(descriptor protoreflect.FieldDescriptor) protor
 		return protoreflect.ValueOfString(value)
 	case "Foo.bar":
 		value := x.Bar
+		return protoreflect.ValueOfMessage(value.ProtoReflect())
+	case "Foo.baz":
+		value := x.Baz
 		return protoreflect.ValueOfMessage(value.ProtoReflect())
 	default:
 		if descriptor.IsExtension() {
@@ -1277,6 +1293,8 @@ func (x *fastReflection_Foo) Set(fd protoreflect.FieldDescriptor, value protoref
 		x.Bro = value.String()
 	case "Foo.bar":
 		x.Bar = value.Message().Interface().(*anypb.Any)
+	case "Foo.baz":
+		x.Baz = value.Message().Interface().(*anypb.Any)
 	default:
 		if fd.IsExtension() {
 			panic(fmt.Errorf("proto3 declared messages do not support extensions: Foo"))
@@ -1302,6 +1320,11 @@ func (x *fastReflection_Foo) Mutable(fd protoreflect.FieldDescriptor) protorefle
 			x.Bar = new(anypb.Any)
 		}
 		return protoreflect.ValueOfMessage(x.Bar.ProtoReflect())
+	case "Foo.baz":
+		if x.Baz == nil {
+			x.Baz = new(anypb.Any)
+		}
+		return protoreflect.ValueOfMessage(x.Baz.ProtoReflect())
 	case "Foo.bro":
 		panic(fmt.Errorf("field bro of message Foo is not mutable"))
 	default:
@@ -1320,6 +1343,9 @@ func (x *fastReflection_Foo) NewField(fd protoreflect.FieldDescriptor) protorefl
 	case "Foo.bro":
 		return protoreflect.ValueOfString("")
 	case "Foo.bar":
+		m := new(anypb.Any)
+		return protoreflect.ValueOfMessage(m.ProtoReflect())
+	case "Foo.baz":
 		m := new(anypb.Any)
 		return protoreflect.ValueOfMessage(m.ProtoReflect())
 	default:
@@ -1701,8 +1727,9 @@ type Foo struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Bro string        `protobuf:"bytes,1,opt,name=bro,proto3" json:"bro,omitempty"`
-	Bar types.Content `protobuf:"bytes,2,opt,name=bar,proto3" json:"bar,omitempty"`
+	Bro string              `protobuf:"bytes,1,opt,name=bro,proto3" json:"bro,omitempty"`
+	Bar types.Content       `protobuf:"bytes,2,opt,name=bar,proto3" json:"bar,omitempty"`
+	Baz authz.Authorization `protobuf:"bytes,3,opt,name=baz,proto3" json:"baz,omitempty"`
 }
 
 func (x *Foo) Reset() {
@@ -1735,6 +1762,13 @@ func (x *Foo) GetBro() string {
 func (x *Foo) GetBar() types.Content {
 	if x != nil {
 		return x.Bar
+	}
+	return nil
+}
+
+func (x *Foo) GetBaz() authz.Authorization {
+	if x != nil {
+		return x.Baz
 	}
 	return nil
 }
@@ -1784,17 +1818,22 @@ var file_testpb_1_proto_rawDesc = []byte{
 	0x6e, 0x75, 0x6d, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x08, 0x4c, 0x49, 0x53, 0x54,
 	0x45, 0x4e, 0x55, 0x4d, 0x42, 0x07, 0x0a, 0x05, 0x4f, 0x4e, 0x45, 0x4f, 0x46, 0x22, 0x11, 0x0a,
 	0x01, 0x42, 0x12, 0x0c, 0x0a, 0x01, 0x78, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x01, 0x78,
-	0x22, 0x57, 0x0a, 0x03, 0x46, 0x6f, 0x6f, 0x12, 0x10, 0x0a, 0x03, 0x62, 0x72, 0x6f, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x62, 0x72, 0x6f, 0x12, 0x3e, 0x0a, 0x03, 0x62, 0x61, 0x72,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x41, 0x6e, 0x79, 0x42, 0x16, 0xca, 0xb4,
-	0x2d, 0x12, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2e, 0x67, 0x6f, 0x76, 0x2e, 0x43, 0x6f, 0x6e,
-	0x74, 0x65, 0x6e, 0x74, 0x52, 0x03, 0x62, 0x61, 0x72, 0x2a, 0x1f, 0x0a, 0x0b, 0x45, 0x6e, 0x75,
-	0x6d, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x07, 0x0a, 0x03, 0x4f, 0x6e, 0x65, 0x10,
-	0x00, 0x12, 0x07, 0x0a, 0x03, 0x54, 0x77, 0x6f, 0x10, 0x01, 0x42, 0x27, 0x5a, 0x25, 0x67, 0x69,
-	0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2f,
-	0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x74, 0x65, 0x73,
-	0x74, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x22, 0x9f, 0x01, 0x0a, 0x03, 0x46, 0x6f, 0x6f, 0x12, 0x10, 0x0a, 0x03, 0x62, 0x72, 0x6f, 0x18,
+	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x62, 0x72, 0x6f, 0x12, 0x3e, 0x0a, 0x03, 0x62, 0x61,
+	0x72, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
+	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x41, 0x6e, 0x79, 0x42, 0x16, 0xca,
+	0xb4, 0x2d, 0x12, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2e, 0x67, 0x6f, 0x76, 0x2e, 0x43, 0x6f,
+	0x6e, 0x74, 0x65, 0x6e, 0x74, 0x52, 0x03, 0x62, 0x61, 0x72, 0x12, 0x46, 0x0a, 0x03, 0x62, 0x61,
+	0x7a, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
+	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x41, 0x6e, 0x79, 0x42, 0x1e, 0xca,
+	0xb4, 0x2d, 0x1a, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2e, 0x61, 0x75, 0x74, 0x68, 0x7a, 0x2e,
+	0x41, 0x75, 0x74, 0x68, 0x6f, 0x72, 0x69, 0x7a, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x03, 0x62,
+	0x61, 0x7a, 0x2a, 0x1f, 0x0a, 0x0b, 0x45, 0x6e, 0x75, 0x6d, 0x65, 0x72, 0x61, 0x74, 0x69, 0x6f,
+	0x6e, 0x12, 0x07, 0x0a, 0x03, 0x4f, 0x6e, 0x65, 0x10, 0x00, 0x12, 0x07, 0x0a, 0x03, 0x54, 0x77,
+	0x6f, 0x10, 0x01, 0x42, 0x27, 0x5a, 0x25, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f,
+	0x6d, 0x2f, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2f, 0x63, 0x6f, 0x73, 0x6d, 0x6f, 0x73, 0x2d,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72,
+	0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -1825,11 +1864,12 @@ var file_testpb_1_proto_depIdxs = []int32{
 	2, // 3: A.ONEOF_B:type_name -> B
 	0, // 4: A.LIST_ENUM:type_name -> Enumeration
 	4, // 5: Foo.bar:type_name -> google.protobuf.Any
-	6, // [6:6] is the sub-list for method output_type
-	6, // [6:6] is the sub-list for method input_type
-	6, // [6:6] is the sub-list for extension type_name
-	6, // [6:6] is the sub-list for extension extendee
-	0, // [0:6] is the sub-list for field type_name
+	4, // 6: Foo.baz:type_name -> google.protobuf.Any
+	7, // [7:7] is the sub-list for method output_type
+	7, // [7:7] is the sub-list for method input_type
+	7, // [7:7] is the sub-list for extension type_name
+	7, // [7:7] is the sub-list for extension extendee
+	0, // [0:7] is the sub-list for field type_name
 }
 
 func init() { file_testpb_1_proto_init() }
